migrated db-common, postgres and two modules from saltcorn-data
 - All db packages are now migrated to typescript.

 - In saltcorn-data "db index" and the "models role" class are migrated.

 - Saltcorn-data runs now from the dist directory.

 - To make it least disruptive, I have tried to maintain
   most of the JavaScript structure with TypeScript features.

 - The tests in db-common and saltcorn-data will execute
   from its dist folder.
