+++ packages/postgres/package.json
@@ -5,12 +5,25 @@
   "homepage": "https://saltcorn.com",
   "scripts": {
     "test": "echo \"Error: no test specified\"",
-    "tsc": "echo \"Error: no TypeScript support yet\"",
-    "clean": "echo \"Error: no TypeScript support yet\""
+    "tsc": "tsc -p tsconfig.json",
+    "clean": "rm -rf ./dist/*"
   },
   "author": "Tom Nielsen",
   "license": "MIT",
-  "main": "index.js",
+  "main": "dist/index.js",
+  "exports": {
+    "./*": "./dist/*.js"
+  },
+  "typesVersions": {
+    "*": {
+      "*": [
+        "dist/*"
+      ]
+    }
+  },
+  "files": [
+    "dist/**/*"
+  ],
   "dependencies": {
     "@saltcorn/db-common": "0.6.2-beta.1",
     "pg": "^8.2.1",
@@ -19,5 +32,11 @@
   "repository": "github:saltcorn/saltcorn",
   "publishConfig": {
     "access": "public"
+  },
+  "devDependencies": {
+    "@types/node": "^16.11.9",
+    "@types/pg-copy-streams": "^1.2.1",
+    "@types/pg": "^8.6.1",
+    "typescript": "^4.4.4"
   }
 }
