+++ packages/saltcorn-data/models/backup.js
@@ -11,7 +11,7 @@ const View = require("./view");
 const File = require("./file");
 const Plugin = require("./plugin");
 const User = require("./user");
-const Role = require("./role");
+const Role = require("./role").default;
 const Page = require("./page");
 const Zip = require("adm-zip");
 const tmp = require("tmp-promise");
