+++ packages/saltcorn-data/models/field.js
@@ -8,7 +8,7 @@
 const db = require("../db");
 const { contract, is } = require("contractis");
 const { recalculate_for_stored, jsexprToWhere } = require("./expression");
-const { sqlsanitize } = require("@saltcorn/db-common/internal.js");
+const { sqlsanitize } = require("@saltcorn/db-common/internal");
 const { InvalidAdminAction } = require("../utils");
 const { mkWhere } = require("../db");
