+++ packages/saltcorn-data/package.json
@@ -4,13 +4,35 @@
   "description": "Data models for Saltcorn, open-source no-code platform",
   "homepage": "https://saltcorn.com",
   "scripts": {
-    "test": "jest --runInBand",
+    "test": "jest dist --runInBand",
     "tsc": "echo \"Error: no TypeScript support yet\"",
     "clean": "echo \"Error: no TypeScript support yet\""
   },
   "author": "Tom Nielsen",
   "license": "MIT",
-  "main": "index.js",
+  "main": "dist/index.js",
+  "exports": {
+    ".": "./dist/index.js",
+    "./*": "./dist/*.js",
+    "./base-plugin/*": "./dist/base-plugin/*.js",
+    "./base-plugin": "./dist/base-plugin/index.js",
+    "./base-plugin/viewtemplate/*": "./dist/base-plugin/viewtemplate/*.js",
+    "./db/": "./dist/db/index.js",
+    "./db": "./dist/db/index.js",
+    "./db/*": "./dist/db/*.js",
+    "./models/*": "./dist/models/*.js"
+  },
+  "typesVersions": {
+    "*": {
+      "*": [
+        "dist/*",
+        "dist/db/*"
+      ]
+    }
+  },
+  "files": [
+    "dist/**/*"
+  ],
   "optionalDependencies": {
     "@saltcorn/postgres": "0.6.2-beta.1",
     "@saltcorn/sqlite": "0.6.2-beta.1"
@@ -51,7 +73,11 @@
   "jest": {
     "testEnvironment": "node",
     "moduleNameMapper": {
-      "@saltcorn/sqlite/(.*)": "@saltcorn/sqlite/dist/$1"
+      "@saltcorn/sqlite/(.*)": "@saltcorn/sqlite/dist/$1",
+      "@saltcorn/postgres/(.*)": "@saltcorn/postgres/dist/$1",
+      "@saltcorn/db-common/(.*)": "@saltcorn/db-common/dist/$1",
+
+      "@saltcorn/data/(.*)": "@saltcorn/data/dist/$1" 
     }
   },
   "publishConfig": {
