+++ packages/saltcorn-data/tests/backup.test.js
@@ -10,7 +10,7 @@ const User = require("../models/user");
 const { setConfig, getConfig } = require("../models/config");
 const Trigger = require("../models/trigger");
 const Library = require("../models/library");
-const Role = require("../models/role");
+const Role = require("../models/role").default;
 afterAll(db.close);
 
 beforeAll(async () => {
