+++ packages/saltcorn-data/tsconfig.ref.json
@@ -1,0 +1,7 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "composite": true,
+    "declaration": true /* set to true by "composite" */
+  },
+}
