+++ packages/server/routes/packs.js
@@ -34,7 +34,7 @@ const {
 const { h5, pre, code, p, text, text_attr } = require("@saltcorn/markup/tags");
 const Library = require("@saltcorn/data/models/library");
 const Trigger = require("@saltcorn/data/models/trigger");
-const Role = require("@saltcorn/data/models/role");
+const Role = require("@saltcorn/data/models/role").default;
 
 /**
  * @type {object}
