+++ packages/server/routes/scapi.js
@@ -15,7 +15,7 @@ const View = require("@saltcorn/data/models/view");
 const Page = require("@saltcorn/data/models/page");
 const File = require("@saltcorn/data/models/file");
 const Trigger = require("@saltcorn/data/models/trigger");
-const Role = require("@saltcorn/data/models/role");
+const Role = require("@saltcorn/data/models/role").default;
 const Tenant = require("@saltcorn/data/models/tenant");
 const Plugin = require("@saltcorn/data/models/plugin");
 const Config = require("@saltcorn/data/models/config");
