+++ tsconfig.json
@@ -2,7 +2,16 @@
   "files": [],
   "references": [
     {
+      "path": "./packages/db-common/tsconfig.ref.json"
+    },
+    {
       "path": "./packages/sqlite/tsconfig.ref.json"
-    }
+    },
+    {
+      "path": "./packages/postgres/tsconfig.ref.json"
+    },
+    {
+      "path": "./packages/saltcorn-data/tsconfig.ref.json"
+    },
   ],
 }
